refactored federation code, added missing activitypub props, added fetching for remote activitypub objects
