+++ Cargo.toml
@@ -80,6 +80,7 @@ http-signing = { git = "https://github.com/lyptt/http-signing.git", rev = "aa354
   "rsa",
 ], default-features = false }
 http = "0.2.8"
+backoff = { version = "0.4.0", features = ["tokio", "futures"] }
 
 [dev-dependencies]
 mockall = "0.11.3"
