+++ src/activitypub/mod.rs
@@ -8,8 +8,11 @@ pub mod helpers;
 pub mod json_ld;
 pub mod key;
 pub mod link;
-pub mod note;
 pub mod object;
+pub mod place;
+pub mod profile;
 pub mod question;
 pub mod rdf_string;
 pub mod reference;
+pub mod relationship;
+pub mod tombstone;
