+++ src/activitypub/tombstone.rs
@@ -1,0 +1,10 @@
+use serde::{Deserialize, Serialize};
+use typed_builder::TypedBuilder;
+
+#[derive(Serialize, Deserialize, PartialEq, Clone, Debug, TypedBuilder)]
+#[builder(field_defaults(default))]
+pub struct TombstoneProps {
+  #[serde(skip_serializing_if = "Option::is_none")]
+  #[serde(rename = "formerType")]
+  pub former_kind: Option<String>,
+}
