+++ src/federation/activitypub/mod.rs
@@ -1,0 +1,6 @@
+mod actor;
+pub mod federate;
+mod note;
+mod util;
+
+pub use federate::*;
